requirements.txt
SQLAlchemy
pandas
numpy
matplotlib
seaborn
dbt
dbt-core
dbt-postgres
python-dotenv
